% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode_kr.R
\name{countrycode_kr}
\alias{countrycode_kr}
\title{Convert Korean-language Country Names into iso3c Country Codes}
\usage{
countrycode_kr(df, country_kr)
}
\description{
Converts Korean country names into iso3c country codes.
Korean governmental data often uses country names only in Korean, and
without universal country codes. Sometimes, Korean-language country names
are not consistent across different datasets. This function is designed to
assign universal iso3c country codes to Korean-language country names that
will make it easier to join different kinds of data.
}
\details{
There can be problems with some countries when used in a country-year
panel data format, especially for countries that have gone through
political transitions such as Germany, Serbia, Vietnam, Yemen and so on.
param df A character vector of source dataframe name.
param country_kr A character vector of the column name that has Korean-language
country names.
}
\examples{
countrycode_kr(df, country_kr)
}
