% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3c.R
\name{iso3c}
\alias{iso3c}
\title{Converting Korean-language Country Names into iso3c Country Codes}
\arguments{
\item{df}{A character vector of source dataframe name.}

\item{country_kr}{A character vector of the column name that has Korean-language
country names.}
}
\value{
\code{iso3c()} returns a dataframe with Korean-language country names.
}
\description{
\code{iso3c()} Converts Korean country names into iso3c country codes.
Korean governmental data often uses country names only in Korean, and
without universal country codes. Sometimes, Korean-language country names
are not consistent across different datasets. This function is designed to
assign universal iso3c country codes to Korean-language country names that
will make it easier to join different kinds of data.

There can be problems with some countries when used in a country-year
panel data format, especially for countries that have gone through
political transitions such as Germany, Serbia, Vietnam, Yemen and so on.
}
\examples{
\dontrun{
df <- iso3c(df, country_kr)
head(df)
}
}
\author{
Kadir Jun Ayhan
}
\keyword{iso3c}
