% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdiplo-package.R
\docType{_PACKAGE}
\name{kdiplo-package}
\alias{kdiplo-package}
\title{kdiplo: A One-Stop Public Repository for Datasets on Korean Diplomacy}
\description{
Generate customized data frames containing various data on Korean
diplomacy and foreign policy.
}
\details{
The \code{kdiplo} package contains functions for generating various data
on Korean diplomacy and foreign policy
}
\references{
\url{https://ayhan.phd/}
\url{https://github.com/kjayhan/kdiplo}
}
\author{
Kadir Jun Ayhan \email{k@ayhan.phd}
}
\keyword{package}
