% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_data}
\alias{trade_data}
\title{Korea's Trade Data}
\format{
A data frame with 12 variables:
\describe{
\item{iso3c}{A character vector of ISO 3166-1 alpha-3 country codes.}
\item{country}{A character vector of country names corresponding to the
iso3c codes.}
\item{year}{A numeric vector of years for which trade data is available.}
\item{export}{A numeric vector representing the USD value of exports from
Korea to the respective countries in each year.
The default is the value reported by the Korean Statistical Information
Service.
It was imputed from the Correlates of War project data when the former was
missing.}
\item{import}{A numeric vector representing the USD value of imports to
Korea from the respective countries in each year.
The default is the value reported by the Korean Statistical Information
Service.It was imputed from the Correlates of War project data when the
former was missing.}
\item{total_export}{A numeric vector indicating the total value of Korea's
exports in each year as reported by Korean Statistical Information Service.}
\item{total_import}{A numeric vector indicating the total value of Korea's
imports in each year as reported by Korean Statistical Information Service.}
\item{export_kosis}{A numeric vector representing export values from Korea
to each country as recorded by the Korean Statistical Information Service.}
\item{import_kosis}{A numeric vector representing import values to Korea
from each country as recorded by the Korean Statistical Information Service.}
\item{export_cow}{A numeric vector representing export values from Korea
to each country as recorded by the Correlates of War project.}
\item{import_cow}{A numeric vector representing import values to Korea
from each country as recorded by the Correlates of War project.}
\item{updated_at}{A date vector indicating the last update date of the data.}
}
}
\source{
\url{https://kosis.kr/statHtml/statHtml.do?orgId=360&tblId=DT_1R11006_FRM101&conn_path=I3}

\url{https://correlatesofwar.org/data-sets/bilateral-trade/}
}
\usage{
trade_data
}
\value{
Returns a dataframe of Korea's export and import in a country-year format.
}
\description{
\code{trade_data} returns a dataset of Korea's export and import
data in a country-year format. It includes detailed information about the trade activities,
such as export and import values by country and year, along with total export and import figures.
}
\keyword{datasets}
